guid: guid! macro ergonomics improvement
With the current structure of the guid! macro, the `uuid` crate must be
added to the dependencies of any crate that uses the macro and in scope.
This provides poor ergonomics for anyone using the macro.

```
error: cannot find macro `uuid` in this scope
   --> my_component\src\component.rs:170:9
    |
170 |         mu_rust_helpers::guid::guid!("4297bd81-1d1d-49aa-b138-7b54e5807264")
    |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    |
    = note: `uuid` is in scope, but it is a crate, not a macro
    = note: this error originates in the macro `mu_rust_helpers::guid::guid` (in Nightly builds, run with -Z macro-backtrace for more info)
help: consider importing this macro
    |
12  + use uuid::uuid;
```

This commit updates the usage of the uuid macro, so that the macro does
not require the `uuid` crate to be in scope. This is done by specifying
that the uuid macro comes from the interior crate (using $crate). This
does, however, re-export the uuid! macro, as it must be public for it
to be used in other crates.